TextTranscribtion/backend/requirements.txt
openai-whisper
yt-dlp
soundfile
librosa
websockets
torch
numpy